Added the PrettyMPIPrinter
This printer is meant to reproduce the vanilla googletest experience as
closely as possible while still providing nice MPI information. It has
to do quite a few things:

 * Make printing only happen on one rank
 * Collect statistics on failed tests as well as which ranks it failed
   on for the summary
 * Make sure tests are marked as failed even if they do not fail on
   rank 0

Writing this is unfortunately a reasonable amount of copy-paste from
googletest/src/gtest.cc due to the fact that googletest doesn't expose
nearly as much of their interface as they should. I also did not take
the Windows ifdefs into account, someone else will have to fix that if
they want it.

Things missing:
 * Windows ifdefs
 * Printing "full test comments", might add this later
 * "Shards", no idea what this is, so I left it out
 * Some of the posix commands might be in .h files, so I should probably
   have use them